FEAT: finish inplementation of SIFT Detector
Rearrange parameters.
Put infrequently modified params in initialization.
Tried to inplement SIFT in two stage.
There are angle and size information in Keypoints.
But there is no function for exacting SIFT descriptors in Python.
Decide to directly use detect and compute if provided.
Also provide a class for manualy defined detector an descriptors.